10/4/2017 Meeting Minutes link
Instead of posting the meeting minutes text as open source, an attempt to be transparent about process but not content, by posting a link to the minutes, only accessible to members with whom a Google Drive folder was shared. Discussion of what is the best practice for posting and sharing meeting minutes will continue at the next meeting.